+++ src/platform/common/variables/customEnvironmentVariablesProvider.node.ts
